Factor out ClearCommand class
