--- src/main/java/seedu/address/logic/commands/appointment/ClearAppointmentCommand.java
@@ -1,4 +1,0 @@
-package seedu.address.logic.commands.appointment;
-
-public class ClearAppointmentCommand {
-}
